+++ apps/manifest-demo/3008-webpack-host/@mf-types/manifest-provider/compiled-types/apps/manifest-demo/3011-rspack-manifest-provider/src/App.d.ts
@@ -1,0 +1,2 @@
+export default App;
+declare function App(): import("react/jsx-runtime").JSX.Element;