+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/runtime
 
+## 0.1.6
+
+### Patch Changes
+
+- 72c7b80: chore: fix release tag
+- Updated dependencies [72c7b80]
+  - @module-federation/sdk@0.1.6
+
+## 0.1.5
+
+### Patch Changes
+
+- 876a4ff: feat: support config shared import:false in runtime
+- f26aa2d: chore: prevent plugins from losing information
+- 1a9c6e7: feat: support config multiple versions shared
+- Updated dependencies [ca271ab]
+- Updated dependencies [1a9c6e7]
+  - @module-federation/sdk@0.1.5
+
 ## 0.1.4
 
 ### Patch Changes